Move x86 architecture specific decoding routines to separate class
