+++ src/boomerang/frontend/CMakeLists.txt
@@ -14,6 +14,7 @@ list(APPEND boomerang-frontend-sources
     frontend/mips/MIPSDecoder
     frontend/mips/MIPSFrontEnd
     frontend/NJMCDecoder
+    frontend/pentium/CapstoneX86Decoder
     frontend/pentium/PentiumFrontEnd
     frontend/pentium/StringInstructionProcessor
     frontend/ppc/PPCDecoder
